f1-service running

+++ Motorsport-Predictor/f1-service/src/main/java/com/motorsport_predictor/f1_service/repositories/ICircuitRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ICircuitRepository extends JpaRepository<Circuit, Integer> {
+public interface ICircuitRepository extends JpaRepository<Circuit, Long> {
 }
