+++ Motorsport-Predictor/f1-service/src/main/java/com/motorsport_predictor/f1_service/repositories/ISprintRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ISprintRepository extends JpaRepository<Sprint, Integer> {
+public interface ISprintRepository extends JpaRepository<Sprint, Long> {
 }
